Add takeScreenshot method with WebElementTester, will take a screenshot of the given element, can specify a padding in pixels to include around the element.
Add takeScreenshot method to RectangleTester to take a screenshot of a given rectangle, can specify padding to include around the rectangle.

Add storeValue method to PointTester and RectangleTester to save element locations to context.

Add RectangleTester.includes(...), it will extend the current rectangle to include the given Rectangle or Point.  Can be used to find the rectangle containing multiple elements to take a cropped screenshot of just those.  Useful for automating screenshots for documentation.